+++ src/TitleDetails/ExpectedPiecesList/ExpectedPiecesList.js
@@ -4,9 +4,9 @@ import PropTypes from 'prop-types';
 import PiecesList from '../PiecesList';
 import { ExpectedPiecesActions } from '../PiecesActions';
 
-const visibleColumns = ['title', 'caption', 'format', 'receiptDate', 'actions'];
+const visibleColumns = ['caption', 'format', 'receiptDate', 'request', 'actions'];
 
-const ExpectedPiecesList = ({ pieces, title, onEditPiece, onReceivePiece }) => {
+const ExpectedPiecesList = ({ pieces, requests, onEditPiece, onReceivePiece }) => {
   const renderActions = (piece) => (
     <ExpectedPiecesActions
       expectedPiece={piece}
@@ -18,8 +18,8 @@ const ExpectedPiecesList = ({ pieces, title, onEditPiece, onReceivePiece }) => {
   return (
     <PiecesList
       pieces={pieces}
+      requests={requests}
       renderActions={renderActions}
-      title={title}
       visibleColumns={visibleColumns}
     />
   );
@@ -29,11 +29,12 @@ ExpectedPiecesList.propTypes = {
   onEditPiece: PropTypes.func.isRequired,
   onReceivePiece: PropTypes.func.isRequired,
   pieces: PropTypes.arrayOf(PropTypes.object),
-  title: PropTypes.string,
+  requests: PropTypes.arrayOf(PropTypes.object),
 };
 
 ExpectedPiecesList.defaultProps = {
   pieces: [],
+  requests: [],
 };
 
 export default ExpectedPiecesList;