+++ src/TitleDetails/PiecesList/PiecesList.test.js
@@ -6,20 +6,28 @@ import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import PiecesList from './PiecesList';
 
-const title = 'ABA Journal';
 const pieces = [{
   caption: 'ABA',
   format: 'Physical',
   receiptDate: '2021-02-06',
   receivedDate: '2020-02-06',
+  itemId: 'item1',
+}];
+const items = [{
+  id: pieces[0].itemId,
+  barcode: 'A43GA',
+}];
+const requests = [{
+  itemId: pieces[0].itemId,
 }];
 
 const renderPiecesList = () => (render(
   <IntlProvider locale="en">
     <PiecesList
       pieces={pieces}
-      title={title}
-      visibleColumns={['title', 'caption', 'format', 'receiptDate', 'receivedDate', 'actions']}
+      items={items}
+      requests={requests}
+      visibleColumns={['barcode', 'caption', 'format', 'receiptDate', 'receivedDate', 'request', 'actions']}
       renderActions={() => 'Actions'}
     />
   </IntlProvider>,
@@ -32,18 +40,20 @@ describe('Given Pieces List', () => {
     const { getByText } = renderPiecesList();
 
     // header is rendered
-    expect(getByText('ui-receiving.piece.title')).toBeDefined();
+    expect(getByText('ui-receiving.piece.barcode')).toBeDefined();
     expect(getByText('ui-receiving.piece.caption')).toBeDefined();
     expect(getByText('ui-receiving.piece.format')).toBeDefined();
     expect(getByText('ui-receiving.piece.receiptDate')).toBeDefined();
     expect(getByText('ui-receiving.piece.receivedDate')).toBeDefined();
+    expect(getByText('ui-receiving.piece.request')).toBeDefined();
 
     // piece item is rendered
-    expect(getByText(title)).toBeDefined();
+    expect(getByText(items[0].barcode)).toBeDefined();
     expect(getByText(pieces[0].caption)).toBeDefined();
-    expect(getByText(pieces[0].format)).toBeDefined();
+    expect(getByText('ui-receiving.piece.pieceFormat.physical')).toBeDefined();
     expect(getByText(pieces[0].receiptDate)).toBeDefined();
     expect(getByText(pieces[0].receivedDate)).toBeDefined();
+    expect(getByText('ui-receiving.piece.request.isOpened')).toBeDefined();
     expect(getByText('Actions')).toBeDefined();
   });
 });