+++ src/TitleDetails/ReceivedPiecesList/ReceivedPiecesList.test.js
@@ -6,10 +6,9 @@ import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import ReceivedPiecesList from './ReceivedPiecesList';
 
-const title = 'ABA Journal';
 const pieces = [{
   caption: 'ABA',
-  format: 'Physical',
+  format: 'Electronic',
   receivedDate: '2020-02-06',
 }];
 
@@ -17,7 +16,6 @@ const renderPiecesList = (onUnreceivePiece) => (render(
   <IntlProvider locale="en">
     <ReceivedPiecesList
       pieces={pieces}
-      title={title}
       onUnreceivePiece={onUnreceivePiece}
     />
   </IntlProvider>,
@@ -36,15 +34,15 @@ describe('Given Received Pieces List', () => {
     const { getByText, getByTestId } = renderPiecesList(onUnreceivePiece);
 
     // header is rendered
-    expect(getByText('ui-receiving.piece.title')).toBeDefined();
+    expect(getByText('ui-receiving.piece.barcode')).toBeDefined();
     expect(getByText('ui-receiving.piece.caption')).toBeDefined();
     expect(getByText('ui-receiving.piece.format')).toBeDefined();
     expect(getByText('ui-receiving.piece.receivedDate')).toBeDefined();
+    expect(getByText('ui-receiving.piece.request')).toBeDefined();
 
     // piece item is rendered
-    expect(getByText(title)).toBeDefined();
     expect(getByText(pieces[0].caption)).toBeDefined();
-    expect(getByText(pieces[0].format)).toBeDefined();
+    expect(getByText('ui-receiving.piece.pieceFormat.electronic')).toBeDefined();
     expect(getByText(pieces[0].receivedDate)).toBeDefined();
     expect(getByTestId('receivedPieceActionMenu')).toBeDefined();
   });