+++ src/TitleDetails/constants.js
@@ -26,6 +26,12 @@ export const PIECE_FORMAT = {
   other: 'Other',
 };
 
+export const PIECE_FORMAT_LABELS = {
+  [PIECE_FORMAT.electronic]: <FormattedMessage id="ui-receiving.piece.pieceFormat.electronic" />,
+  [PIECE_FORMAT.physical]: <FormattedMessage id="ui-receiving.piece.pieceFormat.physical" />,
+  [PIECE_FORMAT.other]: <FormattedMessage id="ui-receiving.piece.pieceFormat.other" />,
+};
+
 export const PIECE_FORMAT_OPTIONS = [
   { labelId: 'ui-receiving.piece.pieceFormat.electronic', value: PIECE_FORMAT.electronic },
   { labelId: 'ui-receiving.piece.pieceFormat.physical', value: PIECE_FORMAT.physical },